Refactor rename scripts so that simplegit import doesn't cause other helper script usages to crash.
