+++ package.json
@@ -38,8 +38,8 @@
     "format": "prettier \"**/*.{js,mjs,ts,jsx,tsx,html,css,scss,json,yml,yaml}\" --write",
     "prepare": "husky install",
     "gi": "node ./scripts/generate/generateIcons.mjs",
-    "rename": "node ./scripts/renameIcon.mjs",
-    "renamePattern": "node ./scripts/renamePattern.mjs"
+    "rename": "node scripts/rename/renameIcon.mjs",
+    "renamePattern": "node scripts/rename/renamePattern.mjs"
   },
   "devDependencies": {
     "@html-eslint/eslint-plugin": "^0.19.1",
