+++ scripts/helpers.mjs
@@ -1,8 +1,6 @@
 import fs from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import simpleGit from 'simple-git';
-import { promisify } from 'util';
 
 /**
  * Converts string to CamelCase
@@ -225,60 +223,4 @@ export function minifySvg(string) {
         .replace(/(<.*?>)|\s+/g, (m, $1) => $1 || ' ')
         .trim()
     : '';
-}
-
-/**
- * Renames an icon and adds the old name as an alias.
- * @param {string} ICONS_DIR
- * @param {string} oldName
- * @param {string} newName
- * @param {boolean} logInfo
- */
-export async function renameIcon(ICONS_DIR, oldName, newName, logInfo = true) {
-  const git = simpleGit();
-
-  async function fileExists(filePath) {
-    try {
-      await promisify(fs.access)(filePath);
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  if (await fileExists(`${ICONS_DIR}/${newName}.svg`)) {
-    throw new Error(`ERROR: Icon icons/${newName}.svg already exists`);
-  }
-  if (await fileExists(`${ICONS_DIR}/${newName}.json`)) {
-    throw new Error(`ERROR: Metadata file icons/${newName}.json already exists`);
-  }
-  if (!(await fileExists(`${ICONS_DIR}/${oldName}.svg`))) {
-    throw new Error(`ERROR: Icon icons/${oldName}.svg doesn't exist`);
-  }
-  if (!(await fileExists(`${ICONS_DIR}/${oldName}.json`))) {
-    throw new Error(`ERROR: Metadata file icons/${oldName}.json doesn't exist`);
-  }
-
-  await git.mv(`${ICONS_DIR}/${oldName}.svg`, `${ICONS_DIR}/${newName}.svg`);
-  await git.mv(`${ICONS_DIR}/${oldName}.json`, `${ICONS_DIR}/${newName}.json`);
-  const json = fs.readFileSync(`${ICONS_DIR}/${newName}.json`, 'utf8');
-  const jsonData = JSON.parse(json);
-  if (Array.isArray(jsonData.aliases)) {
-    jsonData.aliases = jsonData.aliases.filter((name) => name !== newName).push(oldName);
-    if (jsonData.aliases.length === 0) {
-      delete jsonData.aliases;
-    }
-  } else {
-    jsonData.aliases = [oldName];
-  }
-  fs.writeFileSync(`${ICONS_DIR}/${newName}.json`, JSON.stringify(jsonData, null, 2));
-  await git.add(`${ICONS_DIR}/${newName}.json`);
-
-  if (logInfo) {
-    console.log('SUCCESS: Next steps:');
-    console.log(`git checkout -b rename/${oldName}-to-${newName};`);
-    console.log(`git commit -m 'Renamed ${oldName} to ${newName}';`);
-    console.log(`gh pr create --title 'Renamed ${oldName} to ${newName}';`);
-    console.log('git checkout main;');
-  }
 }
